requirements.txt
gensim==3.8.3
matplotlib==3.4.2
numpy==1.21
pandas==1.3.0
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.7.0
seaborn==0.11.1
opencv-python==4.5.1.48
nltk==3.8.1
biopython==1.80
skfeature-chappers==1.1.0
#word2vec==0.11.1
#Cython==0.29.21
#ipykernel==5.3.4